refactor(core): blog fragment colocation
