+++ apps/core/app/[locale]/(default)/blog/tag/[tagId]/page.tsx
@@ -2,9 +2,10 @@ import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { notFound } from 'next/navigation';
 import { getTranslations } from 'next-intl/server';
 
-import { getBlogPosts } from '~/client/queries/get-blog-posts';
 import { BlogPostCard } from '~/components/blog-post-card';
 import { Link } from '~/components/link';
+
+import { getBlogPosts } from '../../page-data';
 
 interface Props {
   params: {
@@ -28,7 +29,7 @@ export default async function BlogPostPage({ params: { tagId, locale }, searchPa
 
       <div className="grid grid-cols-1 gap-10 sm:grid-cols-2 lg:grid-cols-3 lg:gap-8">
         {blogPosts.posts.items.map((post) => {
-          return <BlogPostCard blogPost={post} key={post.entityId} />;
+          return <BlogPostCard data={post} key={post.entityId} />;
         })}
       </div>
